+++ frontend/components/Sheet/index.spec.jsx
@@ -16,6 +16,10 @@ jest.mock('../../selectors', () => ({
 const MockedGridComponent = () => <div>Hello world</div>;
 jest.mock('../../components/Grid', () => MockedGridComponent);
 
+jest.mock('@shopgate/pwa-extension-kit/env/helpers', () => ({
+  isIOSTheme: () => false,
+}));
+
 describe('Sheet', () => {
   // eslint-disable-next-line global-require
   const Sheet = require('./index').default;